Remove SingletonLazyTest and AbstractMultithreadedLazyTest
* SingletonLazyTest is covered by MultithreadedLazyTest
* As there is only one multithreaded test now, no need in extracting
  common code into abstract class
