+++ src/test/java/ru/spbau/mit/MultithreadedLazyTest.java
@@ -5,15 +5,18 @@ import ru.spbau.mit.utils.TestSupplier;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
+import ru.spbau.mit.utils.TestThread;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.function.Supplier;
 
 import static org.junit.Assert.*;
 
 @RunWith(Parameterized.class)
-public class MultithreadedLazyTest extends AbstractMultithreadedLazyTest {
+public class MultithreadedLazyTest {
     @Parameterized.Parameters
     public static Collection<Object[]> data() {
         return Arrays.asList(new Object[][] {
@@ -30,12 +33,54 @@ public class MultithreadedLazyTest extends AbstractMultithreadedLazyTest {
         this.shouldHaveConflicts = shouldHaveConflicts;
     }
 
+    private final int THREADS_COUNT = 2000;
+    private final int LAZIES_COUNT = 200;
+
+    private Integer[][] runThreads(TestSupplier[] suppliers) throws InterruptedException {
+        final TestThread[] ths = new TestThread[THREADS_COUNT];
+
+        final ArrayList<Lazy<Integer>> lazies = new ArrayList<>();
+        final Integer[][] results = new Integer[ths.length][LAZIES_COUNT];
+
+        for (TestSupplier supplier : suppliers) {
+            lazies.add(factory.createLazy(supplier));
+        }
+
+        for (int i = 0; i < ths.length; i++) {
+            final int id = i;
+            ths[i] = new TestThread() {
+                @Override
+                public void run() throws Exception {
+                    ArrayList<Integer> order = new ArrayList<Integer>();
+                    for (int i = 0; i < LAZIES_COUNT; i++) {
+                        order.add(i);
+                    }
+                    Collections.shuffle(order);
+                    for (Integer lazy : order) {
+                        results[id][lazy] = lazies.get(lazy).get();
+                    }
+                }
+            };
+        }
+
+        for (TestSupplier s : suppliers) {
+            assertEquals(0, s.getCallsCount());
+        }
+        for (TestThread th : ths) {
+            th.start();
+        }
+        for (TestThread th : ths) {
+            th.join();
+        }
+        return results;
+    }
+
     private void test(Supplier<TestSupplier> supplierFactory, boolean shouldHaveNull) throws InterruptedException {
         final TestSupplier[] suppliers = new TestSupplier[LAZIES_COUNT];
         for (int i = 0; i < suppliers.length; i++) {
             suppliers[i] = supplierFactory.get();
         }
-        Integer[][] results = performTest(factory, suppliers);
+        Integer[][] results = runThreads(suppliers);
 
         boolean haveNull = false;
         boolean haveNonNull = false;
